requirements.txt
meshtastic[cli]
pyserial
pypubsub
tabulate
protobuf
rich
paho-mqtt
python-dotenv
